Move checkPathForErrors to forms-shared, add basic tests
